raw_rest.rb: A helper class for debugging API responses.
The ShefRawREST class can be used to create a raw_api object that
behaves similarly to Shef's api object but shows the JSON actually
returned by the API rather than inflating the JSON into the relevant
Chef object.  This is helpful for debugging problems with the API
which are hidden by the inflation process.